Consolidate file doc blocks sniffs.
