+++ Spryker/Sniffs/Commenting/AbstractFileDocBlockSniff.php
@@ -41,7 +41,6 @@ abstract class AbstractFileDocBlockSniff extends AbstractSprykerSniff
      */
     protected $licenseMap = [];
 
-
     /**
      * @return array
      */
@@ -157,6 +156,97 @@ abstract class AbstractFileDocBlockSniff extends AbstractSprykerSniff
             $phpCsFile->fixer->replaceToken($currentPosition, '');
             $currentPosition++;
         }
+    }
+
+    /**
+     * @param \PHP_CodeSniffer\Files\File $phpCsFile
+     * @param int $stackPointer
+     *
+     * @return void
+     */
+    protected function checkCustomFileDocBlock(File $phpCsFile, $stackPointer)
+    {
+        $file = getcwd() . DIRECTORY_SEPARATOR . '.license';
+        $license = $this->getLicense($file);
+        if (!$license) {
+            return;
+        }
+
+        if (!$this->existsFileDocBlock($phpCsFile, $stackPointer)) {
+            $fix = $phpCsFile->addFixableError(basename($phpCsFile->getFilename()) . ' has no File Doc Block.', $stackPointer, 'CustomFileDocBlockMissing');
+            if ($fix) {
+                $this->addFileDocBlock($phpCsFile, 0);
+            }
+            return;
+        }
+
+        if ($this->isCustomFileDocBlock($phpCsFile, $stackPointer, $license)) {
+            return;
+        }
+
+        $fix = $phpCsFile->addFixableError(basename($phpCsFile->getFilename()) . ' has the wrong file doc block', $stackPointer, 'CustomFileDocBlockWrong');
+        if ($fix) {
+            $this->addCustomFileDocBlock($phpCsFile, 0, $license);
+        }
+    }
+
+    /**
+     * @param \PHP_CodeSniffer\Files\File $phpCsFile
+     * @param int $stackPointer
+     * @param string $license
+     *
+     * @return bool
+     */
+    protected function isCustomFileDocBlock(File $phpCsFile, $stackPointer, $license)
+    {
+        $fileDockBlockTokens = $this->getFileDocBlockTokens($phpCsFile, $stackPointer);
+
+        $comment = '';
+        foreach ($fileDockBlockTokens as $fileDockBlockToken) {
+            $comment .= $fileDockBlockToken['content'];
+        }
+
+        if (trim($comment) === trim($license)) {
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * @param \PHP_CodeSniffer\Files\File $phpCsFile
+     * @param int $stackPointer
+     *
+     * @return array
+     */
+    protected function getFileDocBlockTokens(File $phpCsFile, $stackPointer)
+    {
+        $fileDocBlockStartPosition = $phpCsFile->findPrevious(T_DOC_COMMENT_OPEN_TAG, $stackPointer);
+        $fileDocBlockEndPosition = $phpCsFile->findNext(T_DOC_COMMENT_CLOSE_TAG, $fileDocBlockStartPosition) + 1;
+
+        $tokens = $phpCsFile->getTokens();
+
+        return array_slice($tokens, $fileDocBlockStartPosition, $fileDocBlockEndPosition - $fileDocBlockStartPosition);
+    }
+
+    /**
+     * @param \PHP_CodeSniffer\Files\File $phpCsFile
+     * @param int $stackPointer
+     * @param string $license
+     *
+     * @return void
+     */
+    protected function addCustomFileDocBlock(File $phpCsFile, $stackPointer, $license)
+    {
+        $phpCsFile->fixer->beginChangeset();
+
+        $this->clearFileDocBlock($phpCsFile, $stackPointer);
+
+        $phpCsFile->fixer->addNewline($stackPointer);
+        $phpCsFile->fixer->addContent($stackPointer, $license);
+        $phpCsFile->fixer->addNewline($stackPointer);
+
+        $phpCsFile->fixer->endChangeset();
     }
 
 }
