--- Spryker/Sniffs/Commenting/DemoshopMissingFileDocBlockSniff.php
@@ -1,44 +1,0 @@
-<?php
-
-namespace Spryker\Sniffs\Commenting;
-
-use PHP_CodeSniffer\Files\File;
-
-/**
- * Checks if file doc block exists
- */
-class DemoshopMissingFileDocBlockSniff extends AbstractDemoshopFileDocBlockSniff
-{
-
-    /**
-     * @param \PHP_CodeSniffer\Files\File $phpCsFile
-     * @param int $stackPointer
-     *
-     * @return void
-     */
-    public function process(File $phpCsFile, $stackPointer)
-    {
-        if (!$this->isPyzNamespace($phpCsFile, $stackPointer)) {
-            return;
-        }
-
-        if (!$this->existsFileDocBlock($phpCsFile, $stackPointer)) {
-            $this->addFixableMissingDocblock($phpCsFile, $stackPointer);
-        }
-    }
-
-    /**
-     * @param \PHP_CodeSniffer\Files\File $phpCsFile
-     * @param int $stackPointer
-     *
-     * @return void
-     */
-    protected function addFixableMissingDocBlock(File $phpCsFile, $stackPointer)
-    {
-        $fix = $phpCsFile->addFixableError(basename($phpCsFile->getFilename()) . ' has no File Doc Block.', $stackPointer, 'FileDocBlockMissing');
-        if ($fix) {
-            $this->addFileDocBlock($phpCsFile, 0);
-        }
-    }
-
-}
