+++ Spryker/ruleset.xml
@@ -110,7 +110,7 @@
     <rule ref="Squiz.WhiteSpace.SemicolonSpacing"/>
     <rule ref="Squiz.WhiteSpace.SuperfluousWhitespace"/>
 
-    <rule ref="Spryker.Commenting.ExistingFileDocBlock">
+    <rule ref="Spryker.Commenting.FileDocBlock">
         <properties>
             <property name="ignorableBundles" type="array" value="Payolution"/>
         </properties>
